porting the java version of unpack to c cleaning up compilation warnings making warning cause errors removing unneeded qdata tool

+++ bsp/qvis3/qvis3.c
@@ -79,14 +79,14 @@ winding_t *NewWinding (int points)
 {
 	winding_t	*w;
 	int			size;
-	
+
 	if (points > MAX_POINTS_ON_WINDING)
 		Error ("NewWinding: %i points", points);
-	
-	size = (int)((winding_t *)0)->points[points];
+
+	size = (intptr_t)((winding_t *)0)->points[points];
 	w = malloc (size);
 	memset (w, 0, size);
-	
+
 	return w;
 }
 
@@ -104,7 +104,7 @@ void prl(leaf_t *l)
 	int			i;
 	portal_t	*p;
 	plane_t		pl;
-	
+
 	for (i=0 ; i<l->numportals ; i++)
 	{
 		p = l->portals[i];
@@ -135,7 +135,7 @@ int PComp (const void *a, const void *b)
 void SortPortals (void)
 {
 	int		i;
-	
+
 	for (i=0 ; i<numportals*2 ; i++)
 		sorted_portals[i] = &portals[i];
 
@@ -213,7 +213,7 @@ void ClusterMerge (int leafnum)
 
 	if (uncompressed[leafnum>>3] & (1<<(leafnum&7)))
 		printf ("WARNING: Leaf portals saw into leaf\n");
-		
+
 	uncompressed[leafnum>>3] |= (1<<(leafnum&7));
 	numvis++;		// count the leaf itself
 
@@ -230,13 +230,13 @@ void ClusterMerge (int leafnum)
 
 	dest = vismap_p;
 	vismap_p += i;
-	
+
 	if (vismap_p > vismap_end)
 		Error ("Vismap expansion overflow");
 
 	dvis->bitofs[leafnum][DVIS_PVS] = dest-vismap;
 
-	memcpy (dest, compressed, i);	
+	memcpy (dest, compressed, i);
 }
 
 
@@ -259,7 +259,7 @@ void CalcPortalVis (void)
 		}
 		return;
 	}
-	
+
 	RunThreadsOnIndividual (numportals*2, true, PortalFlow);
 
 }
@@ -279,7 +279,7 @@ void CalcVis (void)
 //	RunThreadsOnIndividual (numportals*2, true, BetterPortalVis);
 
 	SortPortals ();
-	
+
 	CalcPortalVis ();
 
 //
@@ -287,7 +287,7 @@ void CalcVis (void)
 //
 	for (i=0 ; i<portalclusters ; i++)
 		ClusterMerge (i);
-		
+
 	printf ("Average clusters visible: %i\n", totalvis / portalclusters);
 }
 
@@ -305,11 +305,11 @@ void SetPortalSphere (portal_t *p)
 	{
 		VectorAdd (total, w->points[i], total);
 	}
-	
+
 	for (i=0 ; i<3 ; i++)
 		total[i] /= w->numpoints;
 
-	bestr = 0;		
+	bestr = 0;
 	for (i=0 ; i<w->numpoints ; i++)
 	{
 		VectorSubtract (w->points[i], total, dist);
@@ -337,7 +337,7 @@ void LoadPortals (char *name)
 	winding_t	*w;
 	int			leafnums[2];
 	plane_t		plane;
-	
+
 	if (!strcmp(name,"-"))
 		f = stdin;
 	else
@@ -358,14 +358,14 @@ void LoadPortals (char *name)
 	// these counts should take advantage of 64 bit systems automatically
 	leafbytes = ((portalclusters+63)&~63)>>3;
 	leaflongs = leafbytes/sizeof(long);
-	
+
 	portalbytes = ((numportals*2+63)&~63)>>3;
 	portallongs = portalbytes/sizeof(long);
 
 // each file portal is split into two memory portals
 	portals = malloc(2*numportals*sizeof(portal_t));
 	memset (portals, 0, 2*numportals*sizeof(portal_t));
-	
+
 	leafs = malloc(portalclusters*sizeof(leaf_t));
 	memset (leafs, 0, portalclusters*sizeof(leaf_t));
 
@@ -377,7 +377,7 @@ void LoadPortals (char *name)
 	vismap_p = (byte *)&dvis->bitofs[portalclusters];
 
 	vismap_end = vismap + MAX_MAP_VISIBILITY;
-		
+
 	for (i=0, p=portals ; i<numportals ; i++)
 	{
 		if (fscanf (f, "%i %i %i ", &numpoints, &leafnums[0], &leafnums[1])
@@ -388,11 +388,11 @@ void LoadPortals (char *name)
 		if ( (unsigned)leafnums[0] > portalclusters
 		|| (unsigned)leafnums[1] > portalclusters)
 			Error ("LoadPortals: reading portal %i", i);
-		
+
 		w = p->winding = NewWinding (numpoints);
 		w->original = true;
 		w->numpoints = numpoints;
-		
+
 		for (j=0 ; j<numpoints ; j++)
 		{
 			double	v[3];
@@ -407,7 +407,7 @@ void LoadPortals (char *name)
 				w->points[j][k] = v[k];
 		}
 		fscanf (f, "\n");
-		
+
 	// calc plane
 		PlaneFromWinding (w, &plane);
 
@@ -417,21 +417,21 @@ void LoadPortals (char *name)
 			Error ("Leaf with too many portals");
 		l->portals[l->numportals] = p;
 		l->numportals++;
-		
+
 		p->winding = w;
 		VectorSubtract (vec3_origin, plane.normal, p->plane.normal);
 		p->plane.dist = -plane.dist;
 		p->leaf = leafnums[1];
 		SetPortalSphere (p);
 		p++;
-		
+
 	// create backwards portal
 		l = &leafs[leafnums[1]];
 		if (l->numportals == MAX_PORTALS_ON_LEAF)
 			Error ("Leaf with too many portals");
 		l->portals[l->numportals] = p;
 		l->numportals++;
-		
+
 		p->winding = NewWinding(w->numpoints);
 		p->winding->numpoints = w->numpoints;
 		for (j=0 ; j<w->numpoints ; j++)
@@ -445,7 +445,7 @@ void LoadPortals (char *name)
 		p++;
 
 	}
-	
+
 	fclose (f);
 }
 
@@ -505,13 +505,13 @@ void CalcPHS (void)
 
 		dest = (long *)vismap_p;
 		vismap_p += j;
-		
+
 		if (vismap_p > vismap_end)
 			Error ("Vismap expansion overflow");
 
 		dvis->bitofs[i][DVIS_PHS] = (byte *)dest-vismap;
 
-		memcpy (dest, compressed, j);	
+		memcpy (dest, compressed, j);
 	}
 
 	printf ("Average clusters hearable: %i\n", count/portalclusters);
@@ -529,7 +529,7 @@ int main (int argc, char **argv)
 	char		name[1024];
 	int		i;
 	double		start, end;
-		
+
 	printf ("---- vis ----\n");
 
 	verbose = false;
@@ -575,10 +575,10 @@ int main (int argc, char **argv)
 		Error ("usage: vis [-threads #] [-level 0-4] [-fast] [-v] bspfile");
 
 	start = I_FloatTime ();
-	
+
 	ThreadSetDefault ();
 
-	SetQdirFromPath (argv[i]);	
+	SetQdirFromPath (argv[i]);
 	strcpy (source, ExpandArg(argv[i]));
 	StripExtension (source);
 	DefaultExtension (source, ".bsp");
@@ -592,21 +592,21 @@ int main (int argc, char **argv)
 	sprintf (portalfile, "%s%s", inbase, ExpandArg(argv[i]));
 	StripExtension (portalfile);
 	strcat (portalfile, ".prt");
-	
+
 	printf ("reading %s\n", portalfile);
 	LoadPortals (portalfile);
-	
+
 	CalcVis ();
 
 	CalcPHS ();
 
-	visdatasize = vismap_p - dvisdata;	
+	visdatasize = vismap_p - dvisdata;
 	printf ("visdatasize:%i  compressed from %i\n", visdatasize, originalvismapsize*2);
 
 	sprintf (name, "%s%s", outbase, source);
 	printf ("writing %s\n", name);
-	WriteBSPFile (name);	
-	
+	WriteBSPFile (name);
+
 	end = I_FloatTime ();
 	printf ("%5.1f seconds elapsed\n", end-start);
